requirements.txt
beautifulsoup4==4.11.1
loguru==0.5.3
Markdown==3.3.4
numpy==1.25.0
pandas==1.5.3
PyMySQL==1.0.2
python_dateutil==2.8.2
python-telegram-bot==21.0.1
SpeechRecognition==3.8.1
todoist_api_python==2.0.2
tabulate==0.9.0
pydub==0.25.1
msal==1.20.0
PyMuPDF==1.21.1
PyGithub==1.58.2
quarter_lib==0.0.22
gradio_client
langchain-groq==0.2.1